Removed dateutil from Pipfile as it no longer exists
The dateutil packages no longer exists and was replaced by python-dateutils. This was causing pipenv install to fail 

Running with the applied changes (log below) allows to VPV to be installed but I'm not sure if it works as the Qt platform plugin "xcb" fails to load on my computing cluster.